requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
frozenlist==1.5.0
google==3.0.0
gunicorn==23.0.0
idna==3.10
multidict==6.4.3
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
twilio==9.5.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
yarl==1.19.0
